requirements.txt
streamlit==1.28.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.17.0
